requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
defusedxml==0.7.1
Django==4.1.2
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
install==1.3.5
Markdown==3.4.1
oauthlib==3.2.2
Pillow==9.2.0
psycopg2==2.9.4
pycparser==2.21
PyJWT==1.7.1
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.5
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sorl-thumbnail==12.9.0
sqlparse==0.4.3
urllib3==1.26.12
